requirements.txt
pyautogen
openai
streamlit
requests
python-dotenv
speechrecognition
pyttsx3
pyaudio
sqlite3